renaming files and organizing directories
